Workouts, Trainer, Sessions fixed
Only error is adding workouts to the sessions

+++ GymWorkoutAPI/DataTransferObjects/GymSessionDto.cs
@@ -3,6 +3,6 @@
 public class GymSessionDto
 {
     public DateTime? SessionDate { get; set; }
-    public float OverallDuration { get; set; }
+    public double OverallDuration { get; set; }
     public int TrainerID { get; set; }
 }
